azure_foundary_hosted_agents-main/agent-framework/agent-with-skills/requirements.txt
azure-ai-agentserver-agentframework>=1.0.0b17
azure-monitor-opentelemetry>=1.6.0
python-frontmatter>=1.1.0
httpx>=0.27.0
